app/requirements.txt
asyncpg
fastapi
uvicorn
SQLAlchemy
python-dotenv
pytest
httpx
